LayoutTests/fast/viewport/viewport-130-expected.txt

ALERT: viewport size 320x352 scale 1 with limits [1, 5] and userScalable true
